Finished save feature. Also restructured the directory structure. Please review the current directory structure and commit.
