+++ Assets/Scripts/Libraries/placeholder.md.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d4bf72f93f764624855b6acc5aee009b
+timeCreated: 1710411036
