+++ Assets/Scripts/Managers/DataManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 19e98024c7534475819586c9c969ca63
+timeCreated: 1709732790
